moving debug.h to include directory
