+++ include/freerdp/debug.h
