+++ libfreerdp-core/secure.c
@@ -23,9 +23,9 @@
 #include "licence.h"
 #include "rdp.h"
 #include "iso.h"
-#include "debug.h"
 #include "tcp.h"
 #include "types.h"
+#include <freerdp/debug.h>
 #include <freerdp/rdpset.h>
 #include <freerdp/utils/memory.h>
